pyproject.toml
[project]
name = "aiciblock-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-nomic>=0.1.0",
    "nomic[local]>=3.0.0",
    "langchain-text-splitters>=1.1.1",
    "openai>=1.0.0",
    "llama-cpp-python>=0.3.0",
    "qdrant-client>=1.12.0",
    "tiktoken>=0.12.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
rerank = [
    "sentence-transformers>=3.0.0",
]
